Breaking out class method implementations
A lot of class method implementations have been declared inline in the header files.
This commit breaks them out into separate cpp files.
Some classes have also been broken into different .h files.
Remove some incorrect comments.
