+++ common/common.cpp
@@ -140,6 +140,185 @@ bool parseInt(const char *arg, int *result)
     *result = value;
     return true;
 }
+FileToolkitWithUndo::~FileToolkitWithUndo()
+{
+    bool success = true;
+    // Clean up all created directories, files, and mount points
+    for (auto it = m_cleanupHandlers.rbegin(); it != m_cleanupHandlers.rend(); ++it) {
+        // Cleaning functions will do their own error/warning output
+        if(isError((*it)->clean())) {
+            success = false;
+        }
+        delete *it;
+    }
+
+    if(!success) {
+        log_error() << "One or more cleanup handlers returned error status, please check the log";
+    }
+}
+
+ReturnCode FileToolkitWithUndo::createParentDirectory(const std::string &path)
+{
+    log_debug() << "Creating parent directories for " << path;
+    std::string parent = parentPath(path);
+    if (!isDirectory(parent) && parent != "") {
+        if(isError(createDirectory(parent))) {
+            log_error() << "Could not create directory " << parent;
+            return ReturnCode::FAILURE;
+        }
+    }
+    return ReturnCode::SUCCESS;
+}
+
+ReturnCode FileToolkitWithUndo::createDirectory(const std::string &path)
+{
+    if (isDirectory(path)) {
+        return ReturnCode::SUCCESS;
+    }
+
+    if(isError(createParentDirectory(path))) {
+        log_error() << "Couldn't create parent directory for " << path;
+        return ReturnCode::FAILURE;
+    }
+
+    if (mkdir(path.c_str(), S_IRWXU | S_IRWXG | S_IRWXO) == -1) {
+        log_error() << "Could not create directory " << path << " - Reason : " << strerror(errno);
+        return ReturnCode::FAILURE;
+    }
+
+    m_cleanupHandlers.push_back(new DirectoryCleanUpHandler(path));
+    log_debug() << "Created directory " << path;
+
+    return ReturnCode::SUCCESS;
+}
+
+ReturnCode FileToolkitWithUndo::bindMount(const std::string &src, const std::string &dst, bool readOnly)
+{
+    unsigned long flags = MS_BIND;
+    const char *fstype = nullptr;
+    const void *data = nullptr;
+    log_debug() << "Bind-mounting " << src << " in " << dst << ", flags: " << flags;
+
+    int mountRes = mount(src.c_str(), dst.c_str(), fstype, flags, data);
+    if (mountRes == 0) {
+        log_verbose() << "Bind-mounted folder " << src << " in " << dst;
+        m_cleanupHandlers.push_back(new MountCleanUpHandler(dst));
+    } else {
+        log_error() << "Could not mount into container: src=" << src
+                    << " , dst=" << dst << " err=" << strerror(errno);
+        return ReturnCode::FAILURE;
+    }
+
+    if (readOnly) {
+        flags = MS_REMOUNT | MS_RDONLY | MS_BIND;
+
+        log_debug() << "Re-mounting read-only" << src << " in "
+                    << dst << ", flags: " << flags;
+        mountRes = mount(src.c_str(), dst.c_str(), fstype, flags, data);
+        if (mountRes != 0) {
+            // Failure
+            log_error() << "Could not re-mount " << src << " , read-only on "
+                        << dst << " err=" << strerror(errno);
+            return ReturnCode::FAILURE;
+        }
+    }
+    return ReturnCode::SUCCESS;
+}
+
+ReturnCode FileToolkitWithUndo::createSharedMountPoint(const std::string &path)
+{
+    // MS_MGC_VAL |
+    auto mountRes = mount(path.c_str(), path.c_str(), "", MS_BIND, nullptr);
+    assert(mountRes == 0);
+    mountRes = mount(path.c_str(), path.c_str(), "", MS_UNBINDABLE, nullptr);
+    assert(mountRes == 0);
+    mountRes = mount(path.c_str(), path.c_str(), "", MS_SHARED, nullptr);
+    assert(mountRes == 0);
+    m_cleanupHandlers.push_back(new MountCleanUpHandler(path));
+    log_debug() << "Created shared mount point at " << path;
+
+    return ReturnCode::SUCCESS;
+}
+
+ReturnCode FileToolkitWithUndo::writeToFile(const std::string &path, const std::string &content)
+{
+    auto ret = softwarecontainer::writeToFile(path, content);
+    if (isError(ret)) {
+        return ret;
+    }
+    m_cleanupHandlers.push_back(new FileCleanUpHandler(path));
+    log_debug() << "Successfully wrote to " << path;
+    return ReturnCode::SUCCESS;
+}
+
+ReturnCode FileToolkitWithUndo::createSymLink(const std::string &source, const std::string &destination)
+{
+    log_debug() << "creating symlink " << source << " pointing to " << destination;
+
+    createDirectory(parentPath(source));
+
+    if (symlink(destination.c_str(), source.c_str()) == 0) {
+        m_cleanupHandlers.push_back(new FileCleanUpHandler(source));
+        log_debug() << "Successfully created symlink from " << source << " to " << destination;
+    } else {
+        log_error() << "Error creating symlink " << destination
+                    << " pointing to " << source << ". Error: "
+                    << strerror(errno);
+        return ReturnCode::FAILURE;
+    }
+    return ReturnCode::SUCCESS;
+}
+
+MountCleanUpHandler::MountCleanUpHandler(const std::string &path)
+{
+    m_path = path;
+}
+
+ReturnCode MountCleanUpHandler::clean()
+{
+    // Lazy unmount. Should be the equivalent of the "umount -l" command.
+    if (umount2(m_path.c_str(), MNT_DETACH) == 0) {
+        log_debug() << "Unmounted " << m_path;
+        return ReturnCode::SUCCESS;
+    } else {
+        log_warn() << "Can't unmount " << m_path << " . Error :" << strerror(errno) << ". Trying to force umount";
+        if (umount2(m_path.c_str(), MNT_FORCE) != 0) {
+            log_warn() << "Can't force unmount " << m_path << " . Error :" << strerror(errno);
+            return ReturnCode::FAILURE;
+        }
+        log_debug() << "Managed to force unmount " << m_path;
+        return ReturnCode::SUCCESS;
+    }
+}
+
+FileCleanUpHandler::FileCleanUpHandler(const std::string &path)
+{
+    m_path = path;
+}
+
+ReturnCode FileCleanUpHandler::clean()
+{
+    if (unlink(m_path.c_str()) == 0) {
+        log_debug() << "Unlinked " << m_path;
+        return ReturnCode::SUCCESS;
+    } else {
+        log_error() << "Can't delete " << m_path << " . Error :" << strerror(errno);
+        return ReturnCode::FAILURE;
+    }
+}
+
+void SignalConnectionsHandler::addConnection(sigc::connection &connection) {
+    m_connections.push_back(connection);
+}
+
+SignalConnectionsHandler::~SignalConnectionsHandler()
+{
+    for (auto &connection : m_connections) {
+        connection.disconnect();
+    }
+}
+
 
 }
+
